+++ bencher/example/example_git_time_event.py
@@ -49,9 +49,8 @@ def example_git_time_event(run_cfg: bn.BenchRunCfg | None = None) -> bn.Bench:
     """
 
     random.seed(42)
-    run_cfg = run_cfg or bn.BenchRunCfg()
+    run_cfg = bn.BenchRunCfg.with_defaults(run_cfg, repeats=5)
     run_cfg.over_time = True
-    run_cfg.repeats = 5
 
     benchable = ServerLatency()
     bench = benchable.to_bench(run_cfg)